Extract helper to read strings from native
