+++ Shared/Realm.Shared/MarshalHelpers.cs
@@ -18,6 +18,7 @@
 
 using System;
 using System.Collections.Generic;
+using System.Diagnostics;
 using System.Runtime.InteropServices;
 
 namespace Realms
@@ -95,6 +96,47 @@ namespace Realms
 
             */
             throw new NotImplementedException("Type " + columnType.FullName + " not supported");
+        }
+
+        public delegate IntPtr NativeStringGetter(IntPtr buffer, IntPtr bufferLength, out bool isNull, out NativeException ex);
+
+        public static string GetString(NativeStringGetter getter)
+        {
+            var bufferSizeNeededChars = 128;
+
+            // First alloc this thread
+            var stringGetBuffer = Marshal.AllocHGlobal((IntPtr)(bufferSizeNeededChars * sizeof(char)));
+            var stringGetBufferLen = bufferSizeNeededChars;
+
+            bool isNull;
+            NativeException nativeException;
+
+            // try to read
+            var bytesRead = (int)getter(stringGetBuffer, (IntPtr)stringGetBufferLen, out isNull, out nativeException);
+            nativeException.ThrowIfNecessary();
+            if (bytesRead == -1)
+            {
+                throw new RealmInvalidDatabaseException("Corrupted string data");
+            }
+
+            if (bytesRead > stringGetBufferLen) // need a bigger buffer
+            {
+                Marshal.FreeHGlobal(stringGetBuffer);
+                stringGetBuffer = Marshal.AllocHGlobal((IntPtr)(bytesRead * sizeof(char)));
+                stringGetBufferLen = bytesRead;
+
+                // try to read with big buffer
+                bytesRead = (int)getter(stringGetBuffer, (IntPtr)stringGetBufferLen, out isNull, out nativeException);
+                nativeException.ThrowIfNecessary();
+                if (bytesRead == -1) // bad UTF-8 in full string
+                {
+                    throw new RealmInvalidDatabaseException("Corrupted string data");
+                }
+
+                Debug.Assert(bytesRead <= stringGetBufferLen, "Buffer must have overflowed.");
+            } // needed re-read with expanded buffer
+
+            return bytesRead != 0 ? Marshal.PtrToStringUni(stringGetBuffer, bytesRead) : (isNull ? null : string.Empty);
         }
     }
 }
